BBV plugin extension for pc and icount
A second gzipped file can be generated containing a line for each BBV
interval giving the start PC, total instruction count when the
interval starts, and length of the interval. Due to the TB granularity
of interval checking, the interval length often slightly exceeds the
targeted length.

The plugin now emits a BBV for the interval in progress at the end of
execution, and a final line with the total instruction count.

Currently an explicit BBV filename has to provided as the first
argument in order to activate this feature,
e.g. "--plugin libbbvgen.so,arg=bbv.gz,arg=bbvi.gz"