Emit lists of top blocks to the info file
When a second argument is provided (the "info" file's name) the plugin
now prints a sorted list of the top 10 blocks in each interval - and a
final dump of the top 10 blocks across the entire execution. The
block's contribution as a percent of the interval (or total execution)
is given at the end of each line.