plugins: add fetch stats to bbvgen
A new plugin argument, "fetch=on" will enable collection of
fetch-centric data. Each BlockInfo has to hold more data in order to
track taken branches and to encode the positions of instruction
boundaries within the block. Also, BlockInfo has to be looked up in
the hash table for every TB execution, which is somewhat
expensive. Overall, enabling the fetch stats adds about 10% in my
small trial case.

The BBV Info file output has new top-level objects: "taken-branches"
with "count" and and a "target-distances" array; and "fetch-stats"
containing a dictionary for each fetch policy (64B and 2x32B for now),
each with "count", "inst-distr", and "bytes-distr" arrays.

The "target-distances" array gives the cumulative count of dynamic
taken branches encountered where "common-bits" LSBs are sufficient to
cover the XOR between the branch and target PCs. For example, the
entry with "common-bits" of 16 gives the cumulative count and overall
percentage of taken branches during execution where the branch and
target were in the same aligned 64K of memory.

The two "fetch-stats" objects give the total number of fetches
required for the particular policy, with the two arrays giving the
distribution of those fetches that covered 1/2/.../33 instructions and
2/4/.../66 bytes.

Resolves #8