plugins: bbv: distance histograms for two-taken
When fetch stats are enabled (fetch=on), two additional histograms are
collected for data about the "first taken" in the "two taken" (2x32B)
fetch policy. These measure the XOR distance between (1) the taken
branch and the target PC and (2) the initial PC of the fetch group and
the target PC.

Distances of less than a cacheline are in bucket "0" and everything
mismatching at bit 15 or above are in bucket "10". The rest are put in
between (e.g. bucket 1 gets things that differ at bit 6, bucket 2 for
differing at bit 7, etc).