rivos: cache plugin: add 'stats' mode
In this mode, a JSON file is generated containing the stats for each
execution region. The new options are 'stats', which takes a path to
the gzipped JSON file to be created, and 'ilen' for setting the
interval instruction count (default: 200M).

The new mode only works for single-core simulations and doesn't emit
top instructions. If the 'stats' parameter isn't set, the plugin
reverts to its former behavior (emitting to the QEMU debug log output
stream if '-d plugins' was set).