RIVOS: target/riscv: support for TEE v0.3
Add config flag and feature, MTTP definitions and CSR. There is no
software-visible control over TEE mode yet, as that is not part of the
architectural definition. Implementations must provide their own TEE
mode enable.

Confidential memory enforcement via MTT is supported in the MMU code.
A target-specific page permission bit "PAGE_CONF" is used to track
whether a TEE-mode request is considered confidential or not (based on
MMU mode, PTE, etc). Translations for non-TEE mode requests are always
checked in the MTT and must be to non-confidential memory.
Confidential requests are flagged as such using the 'secure' bit in
the MemTxAttrs, which is kept in the QEMU TLB entry for the page and
is sent along with each MMIO transaction to non-RAM regions.

If necessary in addition to the MemTxAttrs secure bit, the CPU can
support a separate AddressSpace for confidential accesses, allowing
for a custom view of the physical address space.

TODO: CSR numbers, NCP PTE bit assignment
TODO: Malformed PTE detection and reporting (Rcode feature?)