RIVOS: target/riscv: add a 'pa-mask' CPU property
Allow the CPU to be given an explicit max physical address size that's
smaller than the default TARGET_PHYS_ADDR_SPACE_BITS value - which is
the max supported by the target architecture.

One use of this will be in CSRs that have bit field widths dictated by
an implementation's actual physical address size.

Signed-off-by: Kip Walker <kip@rivosinc.com>