Add support for RISC-V QOS identifiers extension
Implement the sqoscfg CSR and tie it to a new privilaged spec extension.
It's used set IDs for resource control and monitoring.
Although the CSR is defined in the S-mode space, the spec says
that accessing it when V=1 shall cause a virtual instruction exception.
Implement this by calling the hmode predicate.