RIVOS: DOE mailbox module
The Rivos implementation of the PCIe Data Object Exchange mailbox uses
a "host" processor to response to messages generated by a
client. There will be multiple instances of DOE mailboxes in Rivos
SoCs, not necessarily appearing with a standard PCIe Extended
Capability Header.

This module implements the client and host register interfaces, the
logic for control and status bits, interrupt delivery, and message
transfer to/from designated buffers in the system. Each DOE Mailbox
instance should be embedded within some QEMU device which will then be
placed into the machine.