rivos: update the CI packaging script
There is no real functional change, the package content is still the
same.

Import the default rivos-sdk template for the Gitlab CI script
and use it to generate the package and simplify the CI script.

Build for the selected Ubuntu architecture and version depending on the
UBUNTU_ARCH and UBUNTU_VERSION CI variables,
but still Ubuntu 20.04 on x86_64 as the default.

Rather than running the builds inside a distro container which is
inside the CI container (Docker in Docker mode), go back to a single
container level by asking the CI to use directly the intended container.