RIVOS: Add RISCVCPUConfig.satp_mode to set sv48, sv57, etc.
RISC-V specifies multiple sizes for addressable memory. It can use for
example 48 or 57 bits to address anything in memory. Linux probes for
the machine's support at startup via the satp CSR register (done in
csr.c:validate_vm).

Because Rivos Sentinel is going to support 48-bits address space, we
want to make sure that QEMU matches it as closely as possible.

This change consists in limiting QEMU to a specific set of satp modes
that can be specified on the command line through the satp_mode CPU
property. It also limits the `-cpu rivos-sentinel` to sv48 by default.

You can set the option as follows:
-cpu riscv,len-satp-mode=3,satp-mode[0]=mbare,satp-mode[1]=sv48,satp-mode[2]=sv57