RIVOS: target/riscv: fix sentinel cpu configuration being overridden by defaults
In rv64_rivos_sentinel_cpu_init(), the call to register_cpu_props() occurs after the sentinel specific parameters are configured. This call overwrites them with defaults, discarding the Sentinel-specific values configured by this function.

Given the repro

int main()
{
    printf(" 8: %d\n", (int)vsetvl_e8m1(256));
    return 0;
}

the expected output when built and executed using qemu-riscv64 -cpu rivos-sentinel is
 8: 32
however, the actual output is
 8: 16

because the default of 128 is used for cpu->cfg.vlen instead of the 256 that rv64_rivos_sentinel_cpu_init() sets. Similar repros can be constructed for the other parameters.

Accordingly, this patch moves the call to register_cpu_props() earlier in the function so it does not undo the work the function performs.