rivos: disable additional builds in CI test pipelines
When the QEMU pipeline is triggered as a downstream pipeline of a longer
CI testing pipeline, execute only one build out of 3 as the goal is just
getting a fresh binary build to do testing (not to release anything).

This should change nothing on regular QEMU pipelines for MR/dev
branches/submission, all builds are still executed/released.